feat: Complete FabricRenderAdapter God Class refactoring and cleanup
Finalize the architecture refactoring by cleaning up old files and completing the v2 migration:

Cleanup Actions:
- Removed old FabricRenderAdapter.ts and IRenderAdapter.ts (v1 files)
- Renamed FabricRenderAdapter.v2.ts → FabricRenderAdapter.ts
- Renamed IRenderAdapter.v2.ts → IRenderAdapter.ts
- Updated FabricRenderAdapterV2 class name → FabricRenderAdapter
- Updated all import references across the codebase

Final Architecture:
- Clean separation of concerns with focused component classes
- Proper delegation pattern implemented (composition over inheritance)
- Legacy compatibility methods preserved for EditorRenderSystem
- All tests passing (345 tests) with clean, maintainable codebase

Key Benefits Achieved:
✅ Single Responsibility Principle - Each class has one clear purpose
✅ Improved testability - Isolated concerns are easier to unit test
✅ Better code reuse - Specialized components can be reused
✅ Easier debugging - Problems are isolated to specific areas
✅ Enhanced maintainability - Changes affect only relevant components

The FabricRenderAdapter God Class has been successfully eliminated and replaced
with a clean, modular architecture following best practices.
